Created an isValidMove function and created a JUnit test for it. Edits made to: Player, TrainCarCards, GameBoard, GameLogic, GameLogicTest.
