--- src/ticketToRide/Card.java
@@ -1,11 +1,0 @@
-package ticketToRide;
-
-// I just created this to make Deck more generic rather than having multiple deck classes.
-public interface Card {
-	
-	//public String toString();
-
-	// Joe's TEST commit
-	// Joe's s SECOND TEST commit
-	
-}
