--- src/ticketToRide/TrainCarCard.java
@@ -1,23 +1,0 @@
-package ticketToRide;
-
-// Add better documentation here. Shouldn't need any additional functionality at this point.
-public class TrainCarCard implements Card{
-	
-	private final String color;
-	
-	public TrainCarCard() {
-		this.color = "null";
-	}
-	
-	public TrainCarCard(String color) {
-		this.color = color;
-	}
-	
-	public String getColor() {
-		return this.color;
-	}
-	
-	public String toString() {
-		return "";
-	}
-}
